requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==21.2.0
newspaper3k==0.2.8
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.12.1
google-cloud-firestore==2.11.1  # New package for Firestore support
pydub==0.25.1
lxml[html_clean]==4.9.3
langdetect==1.0.9
beautifulsoup4==4.12.2
requests==2.31.0